Revert "Creating mapping"
This reverts commit 08edf79b
