+++ fakturogen-invoice/src/main/java/pl/fakturogen/invoice/web/dto/CustomerDTO.java
@@ -1,11 +1,4 @@
 package pl.fakturogen.invoice.web.dto;
 
-import lombok.Data;
-
-@Data
 public class CustomerDTO {
-
-    private Long id;
-    private String name;
-
 }
