+++ fakturogen-invoice/src/main/java/pl/fakturogen/invoice/web/dto/InvoiceSaveDTO.java
@@ -2,7 +2,6 @@ package pl.fakturogen.invoice.web.dto;
 
 import lombok.Builder;
 import lombok.Data;
-import org.springframework.format.annotation.DateTimeFormat;
 import pl.fakturogen.invoice.dao.entity.Product;
 
 import java.time.LocalDate;
@@ -13,17 +12,16 @@ import java.util.List;
 public class InvoiceSaveDTO {
 
     private String number;
-    @DateTimeFormat(pattern = "dd-MM-YYYY")
     private LocalDate issueDate;
     private LocalDate saleDate;
     private LocalDate dueDate;
-    private Integer paymentMethod;
+    private String paymentMethod;
     private double total;
     private double tax;
     private double net;
     private double discount;
     private Integer status;
-    private CustomerDTO customerDTO;
+    private Integer customerId;
     private List<Product> items;
     private Integer bankAccountId;
     private Integer invoiceType;
